refactor: Move infrastructure exceptions to infrastructure/exceptions.py
No error data classes exist in this layer so errors.py is removed entirely.
