+++ sharedkernel/infrastructure/__init__.py
@@ -1,5 +1,6 @@
-from . import services
+from . import exceptions, services
 
 __all__ = [
+    "exceptions",
     "services",
 ]
